+++ src/lib/workflows/audit/index.ts
@@ -1,11 +1,7 @@
-import { spawn } from 'node:child_process';
-import { existsSync } from 'node:fs';
-import { join } from 'node:path';
 import {
   AGENT_SKILL_STEPS,
   createSkillWorkflow,
 } from '../agent-skill/index.js';
-import { OutroKind } from '../../wizard-session.js';
 import type { Workflow, WorkflowConfig } from '../workflow-step.js';
 import { AUDIT_ABORT_CASES } from './detect.js';
 import {
@@ -15,29 +11,6 @@ import {
 } from './types.js';
 
 const REPORT_FILE = 'posthog-audit-report.md';
-
-const PLATFORM_OPEN_CMD: Record<NodeJS.Platform, string> = {
-  darwin: 'open',
-  win32: 'start',
-  aix: 'xdg-open',
-  android: 'xdg-open',
-  freebsd: 'xdg-open',
-  haiku: 'xdg-open',
-  linux: 'xdg-open',
-  openbsd: 'xdg-open',
-  sunos: 'xdg-open',
-  netbsd: 'xdg-open',
-  cygwin: 'xdg-open',
-};
-
-function openReport(absolutePath: string): void {
-  const cmd = PLATFORM_OPEN_CMD[process.platform] ?? 'xdg-open';
-  try {
-    spawn(cmd, [absolutePath], { detached: true, stdio: 'ignore' }).unref();
-  } catch {
-    // Best-effort; user can still open the file manually.
-  }
-}
 
 /** Audit reuses the agent-skill step pipeline but swaps in audit-specific
  *  screens for intro / run / outro. The screen names are registered in
@@ -68,20 +41,6 @@ const baseConfig = createSkillWorkflow({
   spinnerMessage: 'Auditing PostHog integration...',
   estimatedDurationMinutes: 5,
   requires: ['posthog-integration'],
-  buildOutroData: (session) => {
-    const reportPath = join(session.installDir, REPORT_FILE);
-    const reportExists = existsSync(reportPath);
-    if (reportExists) openReport(reportPath);
-    return {
-      kind: OutroKind.Success,
-      message: 'Audit complete!',
-      body: reportExists
-        ? `Report saved to ${reportPath} — opening it now.`
-        : `Report should have been written to ${reportPath} but the file is missing.`,
-      reportFile: reportExists ? REPORT_FILE : undefined,
-      docsUrl: 'https://posthog.com/docs/product-analytics/best-practices',
-    };
-  },
   fileWatchers: [
     {
       filename: AUDIT_CHECKS_FILE,
